requirements.txt
asgiref==3.2.3
certifi==2019.11.28
dj-database-url==0.5.0
dj-email-url==0.2.0
Django==3.0.2
gunicorn==20.0.4
Pillow==7.0.0
psycopg2-binary==2.8.4
python-dotenv==0.10.5
pytz==2019.3
selenium==3.141.0
sentry-sdk==0.14.1
sqlparse==0.3.0
urllib3==1.25.8
whitenoise==5.0.1
